# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning Platform API",
        "description": "API documentation for the Online Learning Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course details",
                "operationId": "bc57e32b05958570eae4745a05618e50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details"
                    }
                }
            }
        },
        "/api/payments/direct": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a direct payment via HyperPay",
                "operationId": "343ed295ea639d81da58f3b3cd5740a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HyperPay response"
                    }
                }
            }
        },
        "/api/payments/result": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Check payment result after 3DS redirect",
                "operationId": "748028b6f900a35f9e5f6ce830fb8a4c",
                "parameters": [
                    {
                        "name": "resourcePath",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to success or failed page"
                    }
                }
            }
        },
...
    },
...
}